Final changes and upload of the Report.

+++ Lab4/src/main/java/UPF_POO20_G101_20/Lab4/users/Buyer.java
@@ -26,7 +26,7 @@ public class Buyer extends User {
 		if (pay(i.getPrice())) {
         	if (i instanceof UnitItem) {
         		UnitItem item = (UnitItem)i;
-        		number = (number > item.getStock()) ? item.getStock() : number;
+        		number = (int)((number > item.getStock()) ? item.getStock() : number);
         	}
         	else if (i instanceof WeightedItem) {
         		WeightedItem item = (WeightedItem)i;
